pyproject.toml
[tool.poetry]
name = "oz-03-collabo-002-be"
version = "0.1.0"
description = ""
authors = ["Gomnonix <lee20481997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
pytest-django = "^4.8.0"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
mypy = "^1.11.1"
django-stubs = "^5.0.4"
djangorestframework-stubs = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
